resin-image-flasher.bbappend: Copy single and quad variant to boot partition
We now suypport both single and quad variants. As these must be used solely on their
designated hardware, we do not default to any of them and copy both to /boot.
Decision on which to be used is to be taken later on, when final image is built and it is
known on which hardware it is supposed to be run.

Signed-off-by: Florin Sarbu <florin@resin.io>