Update meta-resin submodule to use branch 1.1.4-ts
We use a separate branch for now because we need latest commits in meta-resin, however meta-resin
has moved passed 1.1.4 release and this new meta-resin codebase is not ready for release. As such,
a new 1.1.4 based branch has been created on top of which there have been added a few commits necessary
for image maker to do some changes needed for the ts boards in a timely manner (without these latest commits,
image maker would take too long to prepare the image to the user, making the user experience bad).

Signed-off-by: Florin Sarbu <florin@resin.io>