systemd_%.bbappend: Do not add the watchdog conf file to the rootfilesystem
This is done temporarily so that the ts4900 will not have a watchdog.
We do this because the watchdog driver used in the ts4900 will not allow
a shutdown, it will just trigger a board boot after 128 seconds. So the
boards stays powered off only for 128 seconds after which it starts
running again.

Signed-off-by: Florin Sarbu <florin@resin.io>