Revert "u-boot-script-ts.bb: Remove recipe"
This reverts commit 6cb43b7388e26872b20e356329e70f3b2293e7ac.

Because the ts4900 board always boots the u-boot from eMMC (even when
booting from the SD card), we need to keep the SD card bootscript in
order to instruct that our root partition is actually /dev/mmcblk1p2.
By default, the board comes pre-installed with an u-boot that is
hardcoded with the root argument set to /dev/mmcblk1p1 and we cannot
go around this without having a bootscript to change this implicit
root parameter.