u-boot-ts_git.bb: Switch to version v2017.09 for ts4900
The switch to Poky Sumo brought in gcc7. With this gcc version
u-boot v2014.10 fails to boot on the ts4900 thus rendering the
board bricked when provisioning or after a HUP. Therefore we
update u-boot to version v2017.09 and also rework all existing
patches for this new codebase.

Changelog-entry: Update u-boot for ts4900 to version v2017.09
Signed-off-by: Florin Sarbu <florin@balena.io>