u-boot-ts_git: Update to newest v2017.09
This update fixes the issue of the board being in a lockup
state at pcie initialization in the kernel after a reboot
(soft reset). Because of the lockup, the watchdog triggers
a reset so the board would be stuck in a reboot loop.

Changelog-entry: Update u-boot in order to fix pcie initialization lockup after reboot(s)
Signed-off-by: Florin Sarbu <florin@balena.io>